services/auth-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
redis==5.0.3
pydantic==2.7.0
pydantic[email]==2.7.0
pydantic-settings==2.2.1
email-validator==2.1.1
httpx==0.27.0

